+++ src/routes/index.js
@@ -1,7 +1,7 @@
 const { Router } = require("express");
 // Importar todos los routers;
 const {products}= require("./products");
-const { carry } = require("./carry");
+const { shoppingList } = require("./shoppingList");
 const { users } = require("./users");
 
 const router = Router();
@@ -12,7 +12,7 @@ router.get("/health", (req, res) => {
 });
 
 // Configurar los routers
-router.use("/carry", carry);
+router.use("/shoppingList", shoppingList);
 router.use("/users", users);
 router.use("/products",products)
